增加菜单管理部分

+++ mall-mbg/src/main/java/com/macro/mall/mapper/SysMenuMapper.java
@@ -10,7 +10,7 @@ public interface SysMenuMapper {
 
     int deleteByExample(SysMenuExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(SysMenu record);
 
@@ -18,7 +18,7 @@ public interface SysMenuMapper {
 
     List<SysMenu> selectByExample(SysMenuExample example);
 
-    SysMenu selectByPrimaryKey(Integer id);
+    SysMenu selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") SysMenu record, @Param("example") SysMenuExample example);
